Extract generateShadowedSourcesJar logic to internal package
